requirements.txt
requests
sqlitedict
pyyaml
fastapi
uvicorn